requirements.txt
babel==2.10.3
python-dateutil==2.8.2
flask-moment==1.0.5
flask-wtf==1.0.1
flask_sqlalchemy==3.0.0
